update: $(date '+%Y-%m-%d %H:%M')

+++ src/api/api.js
@@ -9,7 +9,9 @@ api.interceptors.request.use((config) => {
   const access_token = localStorage.getItem("access_token");
   if (access_token) {
     config.headers.Authorization = `Bearer ${access_token}`;
+    console.log("accessinng");
   }
+  config.headers["ngrok-skip-browser-warning"] = "true";
   return config;
 });
 
@@ -18,10 +20,16 @@ api.interceptors.response.use(
   async (error) => {
     const originalRequest = error.config;
 
+    if (originalRequest.url.includes("auth/login/admin")) {
+      return Promise.reject(error);
+    }
+
+    console.log("refreshing");
+
     if (error.response?.status === 401 && !originalRequest._retry) {
       originalRequest._retry = true;
-      const refresh_token = localStorage.getItem("refresh_token");
 
+      const refresh_token = localStorage.getItem("refresh_token");
       if (!refresh_token) {
         localStorage.clear();
         window.location.href = "/signin";
@@ -33,15 +41,19 @@ api.interceptors.response.use(
           `${BASE_URL}/auth/refresh/admin`,
           {},
           {
-            headers: { Authorization: `Bearer ${refresh_token}` },
+            headers: {
+              Authorization: `Bearer ${refresh_token}`,
+              "ngrok-skip-browser-warning": "true",
+            },
           }
         );
 
-        const newAccess = data.access_token;
+        const newAccess = data.data.access_token;
         localStorage.setItem("access_token", newAccess);
         originalRequest.headers.Authorization = `Bearer ${newAccess}`;
         return api(originalRequest);
       } catch (err) {
+        console.log("error");
         localStorage.clear();
         window.location.href = "/signin";
         return Promise.reject(err);
